chore: 로그인 관련 오류 수정 refactor: 닉네임 설정 관련 리팩토링
