+++ src/api/user-action.ts
@@ -29,7 +29,7 @@ export const updateNickname = async ({
   const supabase = createClient();
   const { error } = await supabase
     .from("user_info")
-    .update({ user_nickname: newNickname })
+    .update({ user_nickname: newNickname, params: { firstTag: true } })
     .eq("user_id", userId)
     .select();
 
@@ -160,19 +160,19 @@ export const getBookmarks = async (
 };
 
 // 회원가입시 params 컬럼
-export const signInParams = async (userId: string) => {
-  const supabase = createClient();
-  const { error } = await supabase
-    .from("user_info")
-    .update({
-      params: { firstTag: false }
-    })
-    .eq("user_id", userId);
+// export const signInParams = async (userId: string) => {
+//   const supabase = createClient();
+//   const { error } = await supabase
+//     .from("user_info")
+//     .update({
+//       params: { firstTag: false }
+//     })
+//     .eq("user_id", userId);
 
-  if (error) {
-    console.error(error);
-  }
-};
+//   if (error) {
+//     console.error(error);
+//   }
+// };
 
 // 닉네임 수정 완료 params 함수
 export const UpdateNicknameParams = async (userId: string) => {
