+++ src/app/page.tsx
@@ -1,13 +1,13 @@
 import Service from "@/components/layout/Service";
-import NicknameManager from "./(auth)/login/components/NicknameManager";
-import { Modal } from "@/components/shared/Modal";
+// import NicknameManager from "./(auth)/login/components/NicknameManager";
+// import { Modal } from "@/components/shared/Modal";
 
 export default function Home() {
   return (
     <>
       <Service />
-      <NicknameManager />
-      <Modal />
+      {/* <NicknameManager />
+      <Modal /> */}
     </>
   );
 }
